scripts/requirements.txt
# For scripts/manage_desktops.py
docker>=6.0.0
psycopg2-binary>=2.9.0
